Set "type":"module" and rename .mjs to .js
