+++ action.yaml
@@ -5,7 +5,7 @@ branding:
   color: 'gray-dark'
 runs:
   using: 'node16'
-  main: 'dist/index.mjs'
+  main: 'dist/index.js'
 inputs:
   confluence_tenant:
     description: 'Your Atlassian tenant name'
