+++ lib/confluence-sdk.js
@@ -2,9 +2,9 @@ import qs from 'node:querystring';
 import fs from 'node:fs';
 import axios from 'axios';
 import FormData from 'form-data';
-import { RequestError } from './confluence-sdk-errors.mjs';
-import util from './util.mjs';
-import logger from './logger.mjs';
+import { RequestError } from './confluence-sdk-errors.js';
+import util from './util.js';
+import logger from './logger.js';
 
 const CONTENT_PATH = '/wiki/rest/api/content';
 const EXPAND_PROPERTIES = [
