+++ lib/confluence-syncer.js
@@ -1,11 +1,11 @@
 import path from 'node:path';
-import context from './context.mjs';
-import config from './config.mjs';
-import logger from './logger.mjs';
-import ConfluenceSDK from './confluence-sdk.mjs';
-import util from './util.mjs';
-import md2html from './md2html.mjs';
-import KrokiSDK from './kroki-sdk.mjs';
+import context from './context.js';
+import config from './config.js';
+import logger from './logger.js';
+import ConfluenceSDK from './confluence-sdk.js';
+import util from './util.js';
+import md2html from './md2html.js';
+import KrokiSDK from './kroki-sdk.js';
 
 const confluence = new ConfluenceSDK(config.confluence);
 const kroki = new KrokiSDK(config.kroki);
