+++ lib/context.js
@@ -1,8 +1,8 @@
 import { readFileSync, existsSync } from 'node:fs';
 import path from 'node:path';
 import YAML from 'yaml';
-import util from './util.mjs';
-import logger from './logger.mjs';
+import util from './util.js';
+import logger from './logger.js';
 
 const MKDOCS_YML = 'mkdocs.yml';
 const README_MD = 'README.md';
