+++ lib/index.js
@@ -1,0 +1,3 @@
+import { sync } from './confluence-syncer.js';
+
+await sync();
