--- lib/index.mjs
@@ -1,3 +1,0 @@
-import { sync } from './confluence-syncer.mjs';
-
-await sync();
