+++ lib/md2html.js
@@ -1,9 +1,9 @@
 import { readFileSync } from 'node:fs';
 import MarkdownIt from 'markdown-it';
-import fence from './plugins/fence.mjs';
-import image from './plugins/image.mjs';
-import link from './plugins/link.mjs';
-import config from './config.mjs';
+import fence from './plugins/fence.js';
+import image from './plugins/image.js';
+import link from './plugins/link.js';
+import config from './config.js';
 
 function getParser() {
     return new MarkdownIt({ xhtmlOut: true, html: true });
