+++ lib/plugins/fence.js
@@ -1,6 +1,6 @@
 import path from 'node:path';
 import { writeFileSync } from 'node:fs';
-import { toConfluenceImage } from './image.mjs';
+import { toConfluenceImage } from './image.js';
 
 export default function plugin(md, options) {
     const supportedGraphs = options.kroki?.enabled && options.kroki?.supportedTypes ? Object.values(options.kroki.supportedTypes) : [];
