+++ lib/plugins/image.js
@@ -1,5 +1,5 @@
 import path from 'node:path';
-import util from '../util.mjs';
+import util from '../util.js';
 
 export default function plugin(md) {
     const _default = md.renderer.rules.image;
