+++ lib/plugins/link.js
@@ -1,4 +1,4 @@
-import util from '../util.mjs';
+import util from '../util.js';
 
 export default function plugin(md) {
     md.renderer.rules.link_open = (tokens, idx, options, env, self) => {
