+++ test/fixtures/sdk_request/update_page/index.js
